Introduced /schema REST path
