+++ src/main/java/com/rackspace/salus/monitor_management/web/controller/MonitorApiController.java
@@ -19,11 +19,9 @@ package com.rackspace.salus.monitor_management.web.controller;
 import static com.rackspace.salus.monitor_management.web.converter.PatchHelper.JSON_PATCH_TYPE;
 
 import com.fasterxml.jackson.annotation.JsonView;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.rackspace.salus.monitor_management.services.MonitorContentTranslationService;
 import com.rackspace.salus.monitor_management.services.MonitorConversionService;
 import com.rackspace.salus.monitor_management.services.MonitorManagement;
-import com.rackspace.salus.monitor_management.services.SchemaService;
 import com.rackspace.salus.monitor_management.web.model.BoundMonitorDTO;
 import com.rackspace.salus.monitor_management.web.model.BoundMonitorsRequest;
 import com.rackspace.salus.monitor_management.web.model.DetailedMonitorInput;
@@ -78,17 +76,14 @@ public class MonitorApiController {
   private MonitorManagement monitorManagement;
   private MonitorConversionService monitorConversionService;
   private final MonitorContentTranslationService monitorContentTranslationService;
-  private final SchemaService schemaService;
 
   @Autowired
   public MonitorApiController(MonitorManagement monitorManagement,
                               MonitorConversionService monitorConversionService,
-                              MonitorContentTranslationService monitorContentTranslationService,
-                              SchemaService schemaService) {
+                              MonitorContentTranslationService monitorContentTranslationService) {
     this.monitorManagement = monitorManagement;
     this.monitorConversionService = monitorConversionService;
     this.monitorContentTranslationService = monitorContentTranslationService;
-    this.schemaService = schemaService;
   }
 
   @GetMapping("/admin/monitors")
@@ -325,11 +320,5 @@ public class MonitorApiController {
   @JsonView(View.Public.class)
   public MultiValueMap<String, String> getMonitorLabelSelectors(@PathVariable String tenantId) {
     return monitorManagement.getTenantMonitorLabelSelectors(tenantId);
-  }
-
-  @GetMapping("/tenant/{tenantId}/monitor-plugins-schema")
-  @JsonView(View.Public.class)
-  public JsonNode getMonitorPluginsSchema() {
-    return schemaService.getMonitorPluginsSchema();
   }
 }
