+++ src/test/resources/MonitorSchemaControllerTest/monitor_plugins_schema_partial.json
